pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ddl-pre-commit-checker"
version = "1.0.0"
description = ""
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3.10",
]
requires-python = "==3.10.*"
dependencies = [
    "pyyaml~=6.0",
    "docker~=6.1.1",
    "psycopg2-binary~=2.9.6",
    "sqlalchemy~=2.0.15"
]

[project.scripts]
check_ddl = "ddl_pre_commit_checker.main:main"